implement generating and saving tags in separate tasks

+++ src/main/java/model/Mp3Model.java
@@ -5,7 +5,9 @@ import main.java.domain.FailedFileDetails;
 import main.java.domain.Mp3FileWrapper;
 import main.java.exceptions.FileNameBadFormatException;
 import main.java.service.Mp3Service;
+import main.java.workers.GenerateTagsWorker;
 import main.java.workers.ImportFilesWorker;
+import main.java.workers.SaveFilesWorker;
 import org.apache.log4j.Logger;
 import org.springframework.stereotype.Component;
 
@@ -133,60 +135,94 @@ public class Mp3Model {
     }
 
     public void generateTagsForImportedFiles(boolean isCyrillicTags) {
+        generateTagsForImportedFiles(isCyrillicTags, null);
+    }
+
+    public void generateTagsForImportedFiles(boolean isCyrillicTags, GenerateTagsWorker worker) {
         lastFailedGeneratingTags = new LinkedList<>();
 
         if (importedFiles.size() > 0) {
             lastModelState = Mp3ModelState.GENRATED;
         }
 
-        for (Mp3File file : importedFiles.values()) {
+        AtomicInteger progress = new AtomicInteger();
+        long maxProgress = importedFiles.size();
+
+        importedFiles.values().parallelStream().forEach(file -> {
             try {
                 ID3v2 id3v2Tag = Mp3Service.crateID3v2Tag(file, isCyrillicTags);
                 file.removeId3v2Tag();
                 file.setId3v2Tag(id3v2Tag);
+                file.setState(Mp3FileWrapper.Mp3FileState.MODIFIED);
             } catch (FileNameBadFormatException e) {
                 logger.debug("Excepton in method generateTagsForImportedFiles: " + e.getMessage());
                 lastFailedGeneratingTags.add(new FailedFileDetails(file.getFilename(), "Failed to generate tags for file '" + file.getFilename() + "'. Reason: " + e.getMessage()));
+                file.setState(Mp3FileWrapper.Mp3FileState.FAILED_MODIFIED);
             } catch (Exception e) {
                 logger.debug("Unexpected excepton in method importFiles: " + e.getMessage());
                 lastFailedGeneratingTags.add(new FailedFileDetails(file.getFilename(), "Unexpected excepton: " + e.getMessage()));
+                file.setState(Mp3FileWrapper.Mp3FileState.FAILED_MODIFIED);
             }
-        }
+
+            if (worker != null) {
+                worker.updateProgress(progress.incrementAndGet(), maxProgress);
+            }
+        });
 
         observers.forEach(Mp3FilesObserver::onTagGenerated);
     }
 
     public void saveImportedFiles() {
+        saveImportedFiles(null);
+    }
+
+    public void saveImportedFiles(SaveFilesWorker worker) {
         lastFailedSavingFiles = new LinkedList<>();
 
         if (importedFiles.size() > 0) {
             lastModelState = Mp3ModelState.SAVED;
         }
 
-        for (Mp3File file : importedFiles.values()) {
+        AtomicInteger progress = new AtomicInteger();
+        long maxProgress = importedFiles.size();
+
+        importedFiles.values().parallelStream().forEach(file -> {
             String originalFileName = file.getFilename();
             String modifiedFileName = originalFileName + ".modified";
+            boolean isSuccess = false;
             try {
-                file.save(modifiedFileName);
+                if (file.getState() != Mp3FileWrapper.Mp3FileState.SAVED && file.getState() != Mp3FileWrapper.Mp3FileState.FAILED_MODIFIED) {
+                    file.save(modifiedFileName);
+                    isSuccess = true;
+                }
             } catch (IOException | NotSupportedException e) {
                 logger.debug("Excepton in method saveImportedFiles: " + e.getMessage());
                 lastFailedSavingFiles.add(new FailedFileDetails(originalFileName, e.getMessage()));
-                continue;
+                file.setState(Mp3FileWrapper.Mp3FileState.FAILED_SAVED);
             } catch (Exception e) {
                 logger.debug("Unexpected excepton in method importFiles: " + e.getMessage());
                 lastFailedSavingFiles.add(new FailedFileDetails(file.getFilename(), "Unexpected excepton: " + e.getMessage()));
-                continue;
+                file.setState(Mp3FileWrapper.Mp3FileState.FAILED_SAVED);
             }
 
-            File oldFile = new File(originalFileName);
-            oldFile.delete();
-            File newlySavedFile = new File(modifiedFileName);
-            boolean isRenameSuccess = newlySavedFile.renameTo(new File(originalFileName));
-            if (!isRenameSuccess) {
-                logger.debug("Excepton in method saveImportedFiles: " + "Failed to rename newly created file. Newly saved file is: " + modifiedFileName);
-                lastFailedSavingFiles.add(new FailedFileDetails(originalFileName, "Failed to rename newly created file. Newly saved file is: " + modifiedFileName));
+            if (isSuccess) {
+                File oldFile = new File(originalFileName);
+                oldFile.delete();
+                File newlySavedFile = new File(modifiedFileName);
+                boolean isRenameSuccess = newlySavedFile.renameTo(new File(originalFileName));
+                if (!isRenameSuccess) {
+                    logger.debug("Excepton in method saveImportedFiles: " + "Failed to rename newly created file. Newly saved file is: " + modifiedFileName);
+                    lastFailedSavingFiles.add(new FailedFileDetails(originalFileName, "Failed to rename newly created file. Newly saved file is: " + modifiedFileName));
+                    file.setState(Mp3FileWrapper.Mp3FileState.FAILED_SAVED);
+                } else {
+                    file.setState(Mp3FileWrapper.Mp3FileState.SAVED);
+                }
             }
-        }
+
+            if (worker != null) {
+                worker.updateProgress(progress.incrementAndGet(), maxProgress);
+            }
+        });
 
         observers.forEach(Mp3FilesObserver::onSavedImportedFiles);
     }
